requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pytest
pytest-asyncio
psycopg[binary]